+++ src/__init__.py
@@ -3,6 +3,6 @@ Scrat-Backup - Plattformübergreifendes Backup-Tool
 Lizenz: GPLv3
 """
 
-__version__ = "0.2.1"
+__version__ = "0.3.0-beta"
 __author__ = "Scrat-Backup Contributors"
 __license__ = "GPLv3"